+++ src/r_interface.c
@@ -78,7 +78,7 @@ extern SEXP R_ggml_opt_dataset_free(SEXP);
 extern SEXP R_ggml_opt_dataset_ndata(SEXP);
 extern SEXP R_ggml_opt_dataset_data(SEXP);
 extern SEXP R_ggml_opt_dataset_labels(SEXP);
-extern SEXP R_ggml_opt_dataset_weights(SEXP);
+extern SEXP R_ggml_opt_dataset_weights(SEXP, SEXP);
 extern SEXP R_ggml_opt_dataset_shuffle(SEXP, SEXP, SEXP);
 extern SEXP R_ggml_opt_dataset_get_batch(SEXP, SEXP, SEXP, SEXP);
 extern SEXP R_ggml_opt_default_params(SEXP, SEXP);
@@ -119,7 +119,7 @@ extern SEXP R_ggml_opt_grad_acc(SEXP, SEXP);
 extern SEXP R_ggml_opt_result_pred(SEXP);
 extern SEXP R_ggml_opt_prepare_alloc(SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP R_ggml_opt_epoch(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
-extern SEXP R_ggml_opt_init_for_fit(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
+extern SEXP R_ggml_opt_init_for_fit(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP R_ggml_opt_init_for_fit_multi(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP R_ggml_sched_sync_threads(SEXP);
 extern SEXP R_ggml_opt_set_lr(SEXP, SEXP, SEXP);
@@ -992,6 +992,7 @@ SEXP R_ggml_diag(SEXP ctx_ptr, SEXP a_ptr);
 
 // Backward pass operations
 SEXP R_ggml_silu_back(SEXP ctx_ptr, SEXP a_ptr, SEXP b_ptr);
+SEXP R_ggml_gelu_back(SEXP ctx_ptr, SEXP a_ptr, SEXP b_ptr);
 SEXP R_ggml_get_rows_back(SEXP ctx_ptr, SEXP a_ptr, SEXP b_ptr, SEXP c_ptr);
 SEXP R_ggml_soft_max_ext_back(SEXP ctx_ptr, SEXP a_ptr, SEXP b_ptr,
                                SEXP scale, SEXP max_bias);
@@ -1544,6 +1545,7 @@ static const R_CallMethodDef CallEntries[] = {
 
     // Backward pass operations
     {"R_ggml_silu_back",                (DL_FUNC) &R_ggml_silu_back,                3},
+    {"R_ggml_gelu_back",                (DL_FUNC) &R_ggml_gelu_back,                3},
     {"R_ggml_get_rows_back",            (DL_FUNC) &R_ggml_get_rows_back,            4},
     {"R_ggml_soft_max_ext_back",        (DL_FUNC) &R_ggml_soft_max_ext_back,        5},
     {"R_ggml_soft_max_ext_back_inplace",(DL_FUNC) &R_ggml_soft_max_ext_back_inplace,5},
@@ -1708,7 +1710,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"R_ggml_opt_dataset_ndata",                (DL_FUNC) &R_ggml_opt_dataset_ndata,                1},
     {"R_ggml_opt_dataset_data",                 (DL_FUNC) &R_ggml_opt_dataset_data,                 1},
     {"R_ggml_opt_dataset_labels",               (DL_FUNC) &R_ggml_opt_dataset_labels,               1},
-    {"R_ggml_opt_dataset_weights",              (DL_FUNC) &R_ggml_opt_dataset_weights,              1},
+    {"R_ggml_opt_dataset_weights",              (DL_FUNC) &R_ggml_opt_dataset_weights,              2},
     {"R_ggml_opt_dataset_shuffle",              (DL_FUNC) &R_ggml_opt_dataset_shuffle,              3},
     {"R_ggml_opt_dataset_get_batch",            (DL_FUNC) &R_ggml_opt_dataset_get_batch,            4},
     {"R_ggml_opt_default_params",               (DL_FUNC) &R_ggml_opt_default_params,               2},
@@ -1749,7 +1751,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"R_ggml_opt_result_pred",                   (DL_FUNC) &R_ggml_opt_result_pred,                   1},
     {"R_ggml_opt_prepare_alloc",                 (DL_FUNC) &R_ggml_opt_prepare_alloc,                 5},
     {"R_ggml_opt_epoch",                         (DL_FUNC) &R_ggml_opt_epoch,                         7},
-    {"R_ggml_opt_init_for_fit",                  (DL_FUNC) &R_ggml_opt_init_for_fit,                  7},
+    {"R_ggml_opt_init_for_fit",                  (DL_FUNC) &R_ggml_opt_init_for_fit,                  8},
     {"R_ggml_opt_init_for_fit_multi",            (DL_FUNC) &R_ggml_opt_init_for_fit_multi,            9},
     {"R_ggml_sched_sync_threads",                (DL_FUNC) &R_ggml_sched_sync_threads,                1},
     {"R_ggml_opt_set_lr",                        (DL_FUNC) &R_ggml_opt_set_lr,                        3},